Revert "Reservation exception"
