+++ src/main/java/com/eathub/api/RestaurantApi.java
@@ -2,9 +2,7 @@ package com.eathub.api;
 
 import com.eathub.conf.SessionConf;
 import com.eathub.dto.OwnerRestaurantDetailDTO;
-
 import com.eathub.dto.RestaurantDetailDTO;
-
 import com.eathub.service.RestaurantService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
-import java.util.List;
 import java.util.Map;
 
 @Slf4j
@@ -75,18 +72,5 @@ public class RestaurantApi {
         return ResponseEntity.ok().body(Map.of(
                 "success", true
         ));
-    }
-    //오늘이라면 오픈시간부터 현재시간 전까지의 timeOption 들을 HHmm 형식으로 반환
-    @PostMapping("/getOutdatedTimes/{restaurantSeq}")
-    public List<String> getOutdatedTimes(@PathVariable("restaurantSeq") Long restaurantSeq) {
-        return restaurantService.getOutdatedTime(restaurantSeq);
-    }
-    //선택한 날짜 중 예약이 된 시간들을 HHmm 형식으로 반환.
-    @PostMapping("/getBookedTimes/{restaurantSeq}/{memberSeq}/{selectedDate}")
-    public List<String> getBookedTimes(@PathVariable("restaurantSeq") Long restaurant_seq,
-                                       @PathVariable("memberSeq") Long memberSeq,
-                                       @PathVariable("selectedDate") String selectedDate
-                                       ) {
-        return restaurantService.getBookedTimes(restaurant_seq, memberSeq, selectedDate);
     }
 }
