+++ src/main/resources/mapper/RestaurantMapper.xml
@@ -91,13 +91,6 @@
         WHERE restaurant_seq = #{restaurant_seq}
     </select>
 
-    <select id="selectOnesReservation" resultType="reservation">
-        SELECT *
-        FROM reservation
-        WHERE restaurant_seq = #{restaurant_seq},
-                    memberSeq = ${member_seq}
-    </select>
-
     <select id="selectRestaurantInfoWithType" resultType="com.eathub.dto.OwnerRestaurantDetailDTO">
         SELECT r.*, c.restaurant_type
         FROM restaurant_info r
