+++ src/main/webapp/WEB-INF/views/members/searchResult.html
@@ -44,6 +44,64 @@
                                 <input class="person" type="hidden" name="person"  th:value="${session.wantingPerson}"/>
                             </a>
                         </div>
+                        <div style="background: rgb(255, 255, 255)">
+                            <div class="speb4g0">
+                                <div class="speb4g1">
+                                    <button class="design_system_pzih8v0" type="button">
+                                        <img src="../../resources/images/map.svg" th:src="@{/images/map.svg}" />
+                                    </button>
+                                    <button class="design_system_gjbv2i0 design_system_gjbv2i1" type="button">
+                                        <img src="../../resources/images/specific.svg" th:src="@{/images/specific.svg}" />
+                                        <span class="design_system_gjbv2i2 design_system_gjbv2i3">1</span>
+                                    </button>
+                                </div>
+                                <span class="speb4g2"></span>
+                                <div class="speb4g3">
+                                    <div
+                                        class="swiper swiper-container swiper-container-initialized swiper-container-horizontal swiper-container-ios swiper-container-backface-hidden"
+                                        id="filterBar"
+                                    >
+                                        <div class="swiper-wrapper">
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4 speb4g5">
+                                                    <span>서울 전체</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide swiper-slide-next">
+                                                <button type="button" class="speb4g4">
+                                                    <span>음식 종류</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4">
+                                                    <span>가격</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4">
+                                                    <span>테이블타입</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4">
+                                                    <span>분위기</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4">
+                                                    <span>편의시설</span>
+                                                </button>
+                                            </div>
+                                            <div class="swiper-slide">
+                                                <button type="button" class="speb4g4">
+                                                    <span>내주변</span>
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
                     </div>
                     <hr class="separator" />
                     <div id="search-result-header" class="_54rkq60">
